goeckoh_cloner/requirements.txt
numpy
scipy
librosa
sounddevice
webrtcvad
openai-whisper
speechbrain
torch
torchaudio
requests
websockets
soundfile
